requirements.txt
beautifulsoup4==4.13.4
fastapi==0.116.1
PyMuPDF
httpx==0.28.1
langchain==0.3.26
langchain_community==0.3.27
langchain_core==0.3.69
langchain_huggingface==0.3.0
newspaper3k
lxml[html_clean]
nltk==3.9.1
openpyxl==3.1.5
pandas==2.3.1
pydantic==2.11.7
python-dotenv==1.1.1
python_docx==1.2.0
Requests==2.32.4
selenium==4.34.2
tweepy==4.16.0
undetected_chromedriver==3.5.5
uvicorn[standard]
sentence-transformers
faiss-cpu